+++ patient-flow/patient-flow-front-end-with-svelte/src/routes/+layout.svelte
@@ -19,6 +19,7 @@
     <a href="/bed-requests">Bed requests</a>
     <a href="/locate">Locate</a>
     <a href="/audits">Audits</a>
+    <a href="/signin">Sign in</a>
   </nav>
 {/if}